Move agent-call file stats out of the schema into the read side
AgentCallFiles.from_call was a from_* constructor doing filesystem I/O
inside schemas.py — both shapes the read-side rule bans. The stat
composition now lives in get_agent_call_files in reads.py and the
schema is declared fields only. Wire shape unchanged.
